[REFACTOR]: 콤마로 구분되어있는 문자열을 나누어 리스트로 반환하는 메소드 PostService로 이동
