+++ 33ma3/src/main/java/softeer/be33ma3/controller/PostController.java
@@ -41,7 +41,11 @@ public class PostController {
     })
     @Operation(summary = "게시글 목록 조회", description = "게시글 목록 조회 메서드 입니다.")
     @GetMapping
-    public ResponseEntity<?> showPosts(@Schema(hidden = true) @CurrentUser Member member) {
+    public ResponseEntity<?> showPosts(@RequestParam(name = "done", required = false) Boolean done,
+                                       @RequestParam(name = "region", required = false) String region,
+                                       @RequestParam(name = "repair", required = false) String repair,
+                                       @RequestParam(name = "tuneUp", required = false) String tuneUp,
+                                       @Schema(hidden = true) @CurrentUser Member member) {
         List<PostThumbnailDto> postThumbnailDtos = postService.showPosts(member);
         return ResponseEntity.ok().body(DataResponse.success("게시글 목록 조회 성공", postThumbnailDtos));
     }
