+++ 33ma3/src/main/java/softeer/be33ma3/service/PostService.java
@@ -175,4 +175,13 @@ public class PostService {
 
         return post;
     }
+
+    // 구분자 콤마로 문자열 파싱 후 각각의 토큰에서 공백 제거 후 리스트 반환
+    public static List<String> stringCommaParsing(String inputString) {
+        if(inputString.isEmpty())
+            return List.of();
+        return Arrays.stream(inputString.split(","))
+                .map(String::strip)
+                .toList();
+    }
 }
